hwt: spe: move to sys/arm64/spe
