+++ sys/conf/files.arm64
@@ -25,10 +25,10 @@ arm64/acpica/OsdEnvironment.c			optional acpi
 arm64/acpica/acpi_wakeup.c			optional acpi
 arm64/acpica/pci_cfgreg.c			optional acpi pci
 arm64/arm64/autoconf.c				standard
-arm64/arm64/arm_spe_backend.c			optional hwt spe
-arm64/arm64/arm_spe_dev.c			optional hwt spe
-arm64/arm64/arm_spe_acpi.c			optional hwt spe acpi
-arm64/arm64/arm_spe_fdt.c			optional hwt spe fdt
+arm64/spe/arm_spe_backend.c			optional hwt spe
+arm64/spe/arm_spe_dev.c				optional hwt spe
+arm64/spe/arm_spe_acpi.c			optional hwt spe acpi
+arm64/spe/arm_spe_fdt.c				optional hwt spe fdt
 arm64/arm64/bus_machdep.c			standard
 arm64/arm64/bus_space_asm.S			standard
 arm64/arm64/busdma_bounce.c			standard
